Separate out api/taskmanager config settings
This commit aims to clean up the issue of api settings being
configured from the taskmanager class, and some neutron settings
not being added to the api.

This fixes the issue where some neutron settings don't get applied
to the api, specifically 'network_label_regex', so the instance
IP address isn't returned to the user.

Change-Id: Iff0849c800846fb12891b2547c26c28e8b6be67f