Separate out api/taskmanager config settings
This commit removes config settings for the API that were set
from within the taskmanager class, and moves them into the API.

This fixes the issue where some neutron settings didn't get
set in the api config file when the api and taskmanager were run on
separate hosts. Specifically, API requests will now return the IP
address of the DB instance when Neutron is used.

Change-Id: I370f137ce12fedb527f1ae7adc013e1dac4dbffe